Fixing int to float problem
from line 55, surface.get_at or maybe the most recent update to python causes the expected int "player["x"]" to become a float, which prevents the game from running. Simply converting them to an int() fixes the problem and the game runs well. If there is a better way to do this I'd really like to know. Still a beginner here.